requirements.txt
instaloader==4.9.6
joblib==1.1.0
numpy==1.23.1
pandas==1.4.3
PyYAML==6.0
scikit_learn==1.2.2
sentence_transformers==2.2.2
spacy==3.5.2
tqdm==4.64.1
xgboost==1.7.5